+++ tests/test_imageMapper.py
@@ -674,7 +674,8 @@ class TestImageMapper(unittest.TestCase):
                 uImage, vImage, *_ = mapper._constructForwardMap(
                     uPupil,
                     vPupil,
-                    inst.getIntrinsicZernikes(*angle, band, jmax=28),
+                    inst.getIntrinsicZernikes(*angle, band),
+                    None,
                     Image(np.zeros((1, 1)), angle, dfType, band),
                 )